+++ packages/components/ui/src/index.js
@@ -1,6 +1,7 @@
 export * from './a11y/index.js';
 export * from './actions/resizeObserver.js';
 export * from './drivers/fonts/index.js';
+export * from './drivers/style/StyleDriver.svelte';
 export * from './icons/index.js';
 // export {default as FetchDriver} from './io/net/FetchDriver.svelte';
 export {default as StorageIO} from './io/storage/StorageIO.svelte';