+++ packages/components/ui/src/utils/url.js
@@ -1,0 +1,8 @@
+const absoluteURLRegex = /^(?:[a-z+]+:)?\/\//iu;
+
+export const getURL = href => new URL(
+	href,
+	absoluteURLRegex.test(href)
+		? undefined
+		: location.origin
+);